requirements.txt
fastapi
uvicorn
python-dotenv
requests
aiofiles
pydantic